imgのメソッドを分割
