+++ inc/glx_img.h
@@ -1,0 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   glx_img.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/17 11:49:22 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/03/17 11:50:29 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef GLX_IMG_H
+# define GLX_IMG_H
+# include "glx.h"
+
+#endif
